requirements.txt
python-docx>=0.8.11
markitdown>=0.1.0
PyQt5>=5.15.0
pandoc>=2.19.2
markdown>=3.8